renamed pages to handlers
